Shellcheck

+++ make.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/env bash
 set -euo pipefail
-unset CDPATH; cd "$( dirname "${BASH_SOURCE[0]}" )"; cd "`pwd -P`"
+unset CDPATH; cd "$( dirname "${BASH_SOURCE[0]}" )"; cd "$(pwd -P)"
 
 # Project settings: package name, test packages (if different), Go & Glide versions, and cross-compilation targets
 pkg="flywheel.io/sdk"
@@ -11,7 +11,7 @@ minGlideV="0.12.3"
 targets=( "linux/amd64" "darwin/amd64" "windows/amd64" )
 #
 
-fatal() { echo -e $1; exit 1; }
+fatal() { echo -e "$1"; exit 1; }
 
 # Check that this project is in a gopath
 test -d ../../../src || fatal "This project must be located in a gopath.\nTry cloning instead to \"src/$pkg\"."
@@ -27,7 +27,8 @@ if [[ "$localOs" == "darwin" ]]; then
 
 	# Load GNU coreutils, findutils, and sed into path
 	suffix="libexec/gnubin"
-	export PATH="$(brew --prefix coreutils)/$suffix:$(brew --prefix findutils)/$suffix:$(brew --prefix gnu-sed)/$suffix:$PATH"
+	export PATH
+	PATH="$(brew --prefix coreutils)/$suffix:$(brew --prefix findutils)/$suffix:$(brew --prefix gnu-sed)/$suffix:$PATH"
 
 	# OSX has shasum. CentOS has sha1sum. Ubuntu has both.
 	alias sha1sum="shasum -a 1"
@@ -47,7 +48,7 @@ prepareGo() {
 	filterError='s/'"I don't have any idea what to do with"'/Download or install failed for go/g;'
 
 	# Install go, clearing tempdir before & after, with nice messaging.
-	test -f $src || (
+	test -f "$src" || (
 		echo "Downloading go $goV..."
 		rm -rf $GIMME_TMP; mkdir -p $GIMME_TMP
 
@@ -59,7 +60,7 @@ prepareGo() {
 	)
 
 	# Load installed go and prepare for compiled tools
-	source $src
+	source "$src"
 	export PATH=$GOPATH/bin:$PATH
 }
 
@@ -76,8 +77,8 @@ prepareGlide() {
 
 	installGlide() {
 		echo "Downloading glide $minGlideV or higher..."
-		mkdir -p $GOPATH/bin
-		rm -f $GOPATH/bin/glide
+		mkdir -p "$GOPATH/bin"
+		rm -f "$GOPATH/bin/glide"
 		curl -sL https://glide.sh/get | bash
 	}
 
@@ -92,7 +93,7 @@ prepareGlide() {
 		generateGlideHash > $glideHashFile
 	}
 
-	test -x $GOPATH/bin/glide || installGlide
+	test -x "$GOPATH/bin/glide" || installGlide
 
 	# Check the current glide version against the minimum project version
 	currentVersion=$(glide --version | cut -f 3 -d ' ' | tr -d 'v')
@@ -105,11 +106,14 @@ prepareGlide() {
 
 	# If glide components are missing, or cache is out of date, run
 	test -f glide.lock -a -d vendor -a -f $glideHashFile || runGlideInstall
-	test `cat $glideHashFile` = `generateGlideHash` || runGlideInstall
+	test "$(cat $glideHashFile)" = "$(generateGlideHash)" || runGlideInstall
 }
 
 build() {
-	extraLdFlags=${1:-}
+	package=${1:-$pkg}
+	extraLdFlags=${2:-}
+
+	# Check glide state, unless this is called from cross-build, which does so once already.
 	if [ -z "$extraLdFlags" ]; then
 		prepareGlide
 	fi
@@ -127,10 +131,11 @@ build() {
 	# The only alternative is *recompiling the standard library N times every build*.
 	# Gimmie, provisioned above, provides a safely writable Go install in your homedir.
 	# https://dave.cheney.net/2015/08/22/cross-compilation-with-go-1-5
-	go install -v -ldflags "-s $extraLdFlags" $pkg 2>&1 | sed "$filterAbsolutePaths $highlightGoFiles"
+	go install -v -ldflags "-s $extraLdFlags" "$package" 2>&1 | sed "$filterAbsolutePaths $highlightGoFiles"
 }
 
 cross() {
+	package=${1:-$pkg}
 	prepareGlide
 
 	# Placed here, instead of at the top, since it's the only place we need it
@@ -149,7 +154,7 @@ cross() {
 		os=${targetSplit[0]}; arch=${targetSplit[1]}
 
 		echo -e "\n-- Building $os $arch --"
-		GOOS=$os GOARCH=$arch build "-X main.BuildHash=$BuildHash -X 'main.BuildDate=$BuildDate'"
+		GOOS=$os GOARCH=$arch build "$package" "-X main.BuildHash=$BuildHash -X 'main.BuildDate=$BuildDate'"
 
 		# Versions of UPX prior to 3.92 had compatibility issues with OSX 10.12 Sierra.
 		# Could augment this with a UPX version check, and compress if local UPX is modern enough.
@@ -164,19 +169,17 @@ cross() {
 			fi
 
 			binary=$( find "$path" -maxdepth 1 | grep -E "${pkg##*/}(\.exe)*" | head -n 1 )
-
-			# which upx > /dev/null && nice upx -q $binary 2>&1 | grep -- "->" || true
+			which upx > /dev/null && nice upx -q "$binary" 2>&1 | grep -- "->" || true
 		fi
 
 		# If this system is the current build target, copy the binary to a build folder.
 		# Makes it easier to export a cross-build.
-
 		if [[ "$localOs" == "$os" && "$arch" =~ .*$localArch ]] ; then
 			path="$GOPATH/bin/"
 			binary=$( find "$path" -maxdepth 1 | grep -E "${pkg##*/}(\.exe)*" | head -n 1 )
 
 			mkdir -p "$GOPATH/bin/${os}_${arch}/"
-			cp $binary "$GOPATH/bin/${os}_${arch}/"
+			cp "$binary" "$GOPATH/bin/${os}_${arch}/"
 		fi
 	done
 
@@ -187,7 +190,7 @@ prepareCrossBuild() {
 	prepareGo
 
 	flag="$GOROOT/.custom-flags/stdlib-cross-compiled"
-	test -f $flag && return 0
+	test -f "$flag" && return 0
 
 	# Filter out packages that are esoteric, require linking, unexported, or outside the stdlib
 	template='{{.ImportPath}}${{.Standard}}'
@@ -203,13 +206,10 @@ prepareCrossBuild() {
 	for package in "${packages[@]}"; do echo "	_ \"$package\"" >> $tempfile; done
 	echo -e ")\nfunc main() { }" >> $tempfile
 
-	(
-		pkg="$pkg/crossCompileStdlib"
-		cross
-		rm -rf $folder
-		find $GOPATH/bin -executable -type f | grep crossCompileStdlib | xargs rm
-		mkdir -p "`dirname "$flag"`"; touch "$flag"
-	)
+	cross "$pkg/crossCompileStdlib"
+	rm -rf $folder
+	find "$GOPATH/bin" -executable -type f | grep "crossCompileStdlib" | xargs rm
+	mkdir -p "$(dirname "$flag")"; touch "$flag"
 }
 
 # Some go tools take package names. Some take file names. Some like the pacakge prefix. Some don't.
@@ -224,13 +224,13 @@ prepareCrossBuild() {
 # List packages with source code: packageA, packageB... First argument is a prefix (optional)
 listPackages() {
 	prefix=${1:-}
-	find . -name '*.go' -printf "%h\n" | sed '/^\.\/vendor/d; s#^\./##g; /\./d; s#^#'$prefix'#g' | sort --unique
+	find . -name '*.go' -printf "%h\n" | sed '/^\.\/vendor/d; s#^\./##g; /\./d; s#^#'"$prefix"'#g' | sort --unique
 }
 
 # List files in the main package. First argument is a prefix (optional)
 listBaseFiles() {
 	prefix=${1:-}
-	find -maxdepth 1 -type f -name '*.go' | sed 's#^\./##g; s#^#'$prefix'#g'
+	find -maxdepth 1 -type f -name '*.go' | sed 's#^\./##g; s#^#'"$prefix"'#g'
 }
 
 format() {
@@ -240,10 +240,10 @@ format() {
 
 formatCheck() {
 	prepareGo
-	badFiles=(`gofmt -l -s $(listPackages && listBaseFiles)`)
+	badFiles=($(gofmt -l -s $(listPackages && listBaseFiles)))
 
 	if [[ ${#badFiles[@]} -gt 0 ]]; then
-		echo "The following files need formatting: " ${badFiles[@]}
+		echo "The following files need formatting: ${badFiles[*]}"
 		exit 1
 	fi
 }
@@ -256,7 +256,7 @@ _test() {
 	# Set which package to test and which package to count coverage against.
 
 	if [[ $testPkg == "" ]]; then
-		go test -v -cover "$@" $pkg $(listPackages $pkg/) 2>&1 | sed -r "$hideEmptyTests"
+		go test -v -cover "$@" "$pkg" $(listPackages $pkg/) 2>&1 | sed -r "$hideEmptyTests"
 	else
 		go test -v -cover -coverprofile=.coverage.out -coverpkg $coverPkg "$@" $testPkg 2>&1 | sed -r "$hideEmptyTests"
 
@@ -267,7 +267,7 @@ _test() {
 clean() {
 	# Remove all build state
 	prepareGo
-	rm -rvf $GOPATH/pkg $GOPATH/bin/${pkg##*/}
+	rm -rvf "$GOPATH/pkg" "${GOPATH:?}/bin/${pkg##*/}"
 }
 
 showEnv() {
@@ -289,7 +289,7 @@ case "$cmd" in
 	"env") # Load environment!   eval $(./make.sh env)
 		showEnv;;
 	*)
-		type ${cmd} >/dev/null 2>&1 && eval ${cmd} || (
+		type "${cmd}" >/dev/null 2>&1 && eval "${cmd}" || (
 			echo "Usage: ./make.sh {go|godoc|gofmt|glide|build|format|clean|test|env|ci|cross}"
 			exit 1
 		);;
